fix: Fix incorrect gobal export alias in buildFlatBarrel
This fixes a bug where an invalid export alias was generated when filenames
had segments that started with a number preceded by a dash or ".".

Examples using dashes and dots:
```typescript
export { default as testFile-1 } from "./directory1/test-file-1";
```
and
```typescript
export { default as testFile.1 } from "./directory1/test.file.1";
```

While underscores are allowed in aliases, dashes and dots is not.

Fix is to extend the regex to also include numbers.